Cambio icon header y rename
