+++ data-detectives-react/src/app/layout.tsx
@@ -4,7 +4,7 @@ import { Inter } from 'next/font/google'
 const inter = Inter({ subsets: ['latin'] })
 
 export const metadata = {
-  title: 'HealthTime',
+  title: 'Health Time',
   description: 'Generated by create next app',
 }
 
@@ -15,6 +15,9 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
+      <head>
+        <link rel="shortcut icon" href="/imgs/favicon.ico" /> 
+      </head>
       <body className={inter.className}>{children}</body>
     </html>
   )
